+++ Source/SwiftLintFramework/Rules/ExplicitInitRule.swift
@@ -78,8 +78,8 @@ public struct ExplicitInitRule: ASTRule, ConfigurationProviderRule, CorrectableR
     }
 
     private func violationRanges(in file: File) -> [NSRange] {
-        return violationRanges(in: file, dictionary: file.structure.dictionary).sorted { lh, rh in
-            lh.location > rh.location
+        return violationRanges(in: file, dictionary: file.structure.dictionary).sorted { lhs, rhs in
+            lhs.location > rhs.location
         }
     }